Server sends ping only when any user is connected. Server checks for inactive users only when any user is connected.
