+++ src/main/java/com/drc/server/service/PingService.java
@@ -1,0 +1,6 @@
+package com.drc.server.service;
+
+public interface PingService {
+
+    void sendPing();
+}
